update base CUDA/OptiX versions to 7.4.0, 11.5. support 32-dword payload size.

+++ tests/run_tests.py
@@ -1,4 +1,5 @@
 import os
+import shutil
 import sys
 import subprocess
 import json
@@ -53,7 +54,8 @@ def run():
             ret = subprocess.run(cmd, check=True)
 
             # RGBAでdiffをとると差が無いことになってしまう。
-            img = Image.open(test['image']).convert('RGB')
+            testImgPath = test['image']
+            img = Image.open(testImgPath).convert('RGB')
             refImg = Image.open(os.path.join(refImgDir, testDir, 'reference.png')).convert('RGB')
             diffImg = ImageChops.difference(img, refImg)
             diffBBox = diffImg.getbbox()
@@ -68,7 +70,10 @@ def run():
             }
 
             for output in test['outputs']:
-                os.remove(output)
+                if config == 'Release' and output == testImgPath and numDiffPixels > 0:
+                    shutil.move(testImgPath, os.path.join(refImgDir, testDir))
+                else:
+                    os.remove(output)
 
             chdir(oldDir)
